add path_utils file
